Safari-609.1.11.1/LayoutTests/webvr/webvr-enabled.html

<!DOCTYPE html>
<script src="../resources/js-test-pre.js"></script>

<script>
if (window.testRunner)
    window.testRunner.dumpAsText();

if (window.internals)
    window.internals.settings.setWebVREnabled(true);

function runTest() {
    if (!window.testRunner)
        return;

    shouldBeDefined(navigator.getVRDisplays);
}

window.addEventListener("load", runTest, false);
</script>
